Move PageWrapper to page folder root
